style(*): format entire source code

+++ protocol/contracts/oracle/Liquidity.sol
@@ -18,8 +18,8 @@ pragma solidity ^0.5.17;
 pragma experimental ABIEncoderV2;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import '@uniswap/v2-core/contracts/interfaces/IUniswapV2Pair.sol';
-import '../external/UniswapV2Library.sol';
+import "@uniswap/v2-core/contracts/interfaces/IUniswapV2Pair.sol";
+import "../external/UniswapV2Library.sol";
 import "../Constants.sol";
 import "./PoolGetters.sol";
 
@@ -28,11 +28,10 @@ contract Liquidity is PoolGetters {
 
     function addLiquidity(uint256 dollarAmount) internal returns (uint256, uint256) {
         (address dollar, address usdc) = (address(dollar()), usdc());
-        (uint reserveA, uint reserveB) = getReserves(dollar, usdc);
+        (uint256 reserveA, uint256 reserveB) = getReserves(dollar, usdc);
 
-        uint256 usdcAmount = (reserveA == 0 && reserveB == 0) ?
-             dollarAmount :
-             UniswapV2Library.quote(dollarAmount, reserveA, reserveB);
+        uint256 usdcAmount =
+            (reserveA == 0 && reserveB == 0) ? dollarAmount : UniswapV2Library.quote(dollarAmount, reserveA, reserveB);
 
         address pair = address(univ2());
         IERC20(dollar).transfer(pair, dollarAmount);
@@ -41,9 +40,10 @@ contract Liquidity is PoolGetters {
     }
 
     // overridable for testing
-    function getReserves(address tokenA, address tokenB) internal view returns (uint reserveA, uint reserveB) {
-        (address token0,) = UniswapV2Library.sortTokens(tokenA, tokenB);
-        (uint reserve0, uint reserve1,) = IUniswapV2Pair(UniswapV2Library.pairFor(UNISWAP_FACTORY, tokenA, tokenB)).getReserves();
+    function getReserves(address tokenA, address tokenB) internal view returns (uint256 reserveA, uint256 reserveB) {
+        (address token0, ) = UniswapV2Library.sortTokens(tokenA, tokenB);
+        (uint256 reserve0, uint256 reserve1, ) =
+            IUniswapV2Pair(UniswapV2Library.pairFor(UNISWAP_FACTORY, tokenA, tokenB)).getReserves();
         (reserveA, reserveB) = tokenA == token0 ? (reserve0, reserve1) : (reserve1, reserve0);
     }
 }
